Refactor back-end models to have one primary function per file
This makes it a lot more manageable and makes the directory easier to navigate, allowing us to see at a glance what model functions there are rather than having to go through one giant file.
