+++ apps/back-end/src/models/blogs/insertBlogStateHistory.ts
@@ -1,0 +1,17 @@
+import type { Connection } from "src/database/connection";
+import type { BlogStateHistoryInsert, BlogStateHistoryRow } from "src/database/schema";
+
+import { blogStateHistoryTable, parseBlogStateHistoryRow } from "src/database/schema";
+
+async function insertBlogStateHistory(
+  connection: Connection,
+  data: BlogStateHistoryInsert,
+): Promise<BlogStateHistoryRow> {
+  const [blogStateHistory] = await connection
+    .insert(blogStateHistoryTable)
+    .values(data)
+    .returning();
+  return parseBlogStateHistoryRow(blogStateHistory);
+}
+
+export default insertBlogStateHistory;
