+++ apps/back-end/src/models/users/insertUser.ts
@@ -1,0 +1,12 @@
+import type { Connection } from "src/database/connection";
+import type { User, UserInsert } from "src/database/schema";
+
+import { assertNotNullable } from "@alextheman/utility";
+
+import { usersTable } from "src/database/schema";
+
+export async function insertUser(connection: Connection, data: UserInsert): Promise<User> {
+  const [user] = await connection.insert(usersTable).values(data).returning();
+  assertNotNullable(user);
+  return user;
+}
