+++ apps/back-end/src/models/users/updateUser.ts
@@ -1,0 +1,22 @@
+import type { Connection } from "src/database/connection";
+import type { User, UserUpdate } from "src/database/schema";
+
+import { eq } from "drizzle-orm";
+
+import { parseUser, usersTable } from "src/database/schema";
+
+async function updateUser(
+  connection: Connection,
+  userId: string,
+  data: UserUpdate,
+): Promise<User | null> {
+  const [user] = await connection
+    .update(usersTable)
+    .set(data)
+    .where(eq(usersTable.id, userId))
+    .returning();
+
+  return user ? parseUser(user) : null;
+}
+
+export default updateUser;
