+++ apps/back-end/src/server/routes/users.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 
 import { getConnection } from "src/database/connection";
-import { selectUser } from "src/models/users";
+import selectUser from "src/models/users/selectUser";
 import handleEndpointMiddleware from "src/utility/handleEndpointMiddleware";
 import resourceNotFoundError from "src/utility/resourceNotFoundError";
 import validateUUID from "src/utility/validators/validateUUID";
@@ -13,7 +13,9 @@ usersRouter
   .route("/:userId")
   .get(
     handleEndpointMiddleware<{ userId: string }>(async (request, response) => {
-      const user = await selectUser(getConnection(), request.params);
+      const connection = getConnection();
+
+      const user = await selectUser(connection, request.params);
 
       if (user === null) {
         throw resourceNotFoundError("user", request.params.userId);
